GUI - seperation of concerns and added pyproject.toml for requirements
