+++ subfix/__main__.py
@@ -1,0 +1,4 @@
+from subfix.gui import launch_gui
+
+if __name__ == "__main__":
+    launch_gui()
